LPD-65357 LPD-67971 portal-search-elasticsearch8-impl: Move pipeline classes to aggregation package
https://liferay.atlassian.net/browse/LPD-67971
