+++ modules/apps/portal-search-elasticsearch8/portal-search-elasticsearch8-impl/src/main/java/com/liferay/portal/search/elasticsearch8/internal/aggregation/ElasticsearchPipelineAggregationTranslator.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.search.elasticsearch8.internal.aggregation.pipeline;
+package com.liferay.portal.search.elasticsearch8.internal.aggregation;
 
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.search.aggregation.pipeline.AvgBucketPipelineAggregation;
